Update copy for assignment creation and course submission
Instructors may not know the next steps after going through the assignment design wizard. This message should make it more clear what to expect.

Also, tweak the warning message for wizard choices that get disabled because the timeline is too short. We still need to revisit the way this info gets presented, per #260.