Update of English message keys and message documentation.
Resolves #25
Resolves #239
Resolves #245
Resolves #246
Resolves #248
Resolves #253

There are a number of message keys that aren't being used and are obsolete. This removes some of them. There are some others that ended up hard-coded after the initial switch to React components, which are updated here but still need to be switched back for i18n keys in the .cjsx files.