Refactor WikiEdits assignment tagging
The main purpose here is to improve the codeclimate score for this class, but it will also resolve (mostly) #228. We want to post assignment templates even if the talk page doesn't exist... but only if the corresponding article does exist.

Before we deploy this, though, we should add a check to make sure only mainspace assignments are being tagged.