Change button label from "Edit Assignments" to "Assign Articles"
This language will be less confusing for instructors.

In order to make that change, I needed to change the way that button labels are generated, which was a "lego" style message built from the word "Edit" plus a Label parameter. Especially for internationalization, but also to keep things greppable and to keep as much of the copy separate from the code as we can, the better way to do this is to avoid procedurally composited interface strings. I think I adjusted all the uses of that label generation code.