[multiline] Fix student revisions drawer opening issue (closes #310)
React gets cranky if the markup you return in your component's render
method is not valid. I was previously returning an empty `<div>` inside
of a table based on the value of @props, which would've caused a <div>
to be rendered as a direct child of a <tbody>, which isn't valid HTML,
which caused a cryptic React Invariant Violation error. Returning an
empty `<tr>` instead fixes it.