Adjust WikiEdits.course_announcement to avoid triggering CAPTCHA
Edits from new users that include external links will trigger a CAPTCHA (and fail). Previously, I tried to work around this by invoking a template that creates the external link to a course, instead of including the link directly in the wikitext. However, this still triggers the CAPTCHA. Including the url as plain text, without the 'http://', should be okay, as that will not get formatted as a link at all. The downside is that users will need to copy and paste the url to reach the course page, but at least the edit should work for new users.