[multiline] Fix ordering of course blocks (closes #312)
Previously, the `order` of all blocks was getting set to `1` because of
the way the form params were being submitted. This commit simplifies the
course form submission (and edit) process with regard to params, and
makes the ordering of course blocks much more explicit. It contains a
migration to fix the ordering of course blocks that have been created
since the last migration that set the order on course blocks.