When a student enrolls, create their sandbox with {{student sandbox}}
This recreates what RagesossBot used to do, in order to preempty the automatic inclusion of a different sandbox template when users first create their sandbox page via the link near their username. That default template has instructions that lead users to the Articles for Creation process, which we want to avoid sending students to.