Add threading for gathering average view data
stats.grok.se only gives us one article's worth of view data per request, so it's the big bottleneck. This moves the logic into ViewImporter (where we aleady had the logic for importing view data for revisions), and it does threading in the same way so that we do up to 30 requests at a time.