[multiline] Fix issue with choosing course start/end dates from overview
My previous commit had the input calling disableSave on the parent
component (overview/details), which changed the state and re-rendered
itself and the child (input element), which make it look like choosing a
different date didn't change the date in the datepicker input. This
commit only calls disableSave when it "should," i.e., if the value of
the input is '' (the value has been blanked).