[ml] change data type of date field to fix apparent timezone issue (closes #278)
Choosing a due date for a block would save the value as the date with
the time of 00:00:00, which Moment then apparently converted to a local
time, which often changed the date when it was displated. For now, we'll
save it as a date unless/until we enable exact times for due dates.