Remove unnecessary argument from Revision#infer_courses_from_user
I was going to change this method to use the Course.current scope instead of searching by start/end date, but then I realized that this doesn't make sense as a Revision instance method. Leaving it as-is ensures that it returns relevant courses even if it's run on a revision from earlier terms. The downside is that for the DYK use case, it won't match courses that just ended where the student is still making edits. But that's a minor issue.