Handle mw-oauth-invalid result from tokens request.
This is the start of fully handling the situation where a user's OAuth credentials expire. It should keep token request errors from propagating through the system and causing other things to fail, and it sets current_user.wiki_token to 'invalid' so that we can handle that situation on the frontend.