[multiline] Set cloned course term to empty string
The spec was to set the cloned course term to '' so that a user would be
able to set it to their desired value. This proved difficult for React
to handle on the front-end because of the interplay of state and props
in text inputs. This commit provides a back-end solution for this
problem by checking the cloned status of the course and setting it to
empty in the JSON response if the course's cloned status is pending
(which it is directly after cloning, as set by the clone manager).