Remove dead code from LookupsController.
The course is not used at all, and the user method doesn't get called from anywhere. These all get called through the Lookup component, which is ultimately called from the cohort_button (for cohort lookup), the tag_button for tag lookup), and the assign_button (for article_lookup). The user lookup is not used anywhere that I can find (and the code does not get hit by the specs).