Fix issue with term on cloned course (Closes #419)
Because of the misspelling of the React component method, the cloned
status of the course was being set to `undefined`, so it was nulling out
the course term in the jbuilder, which caused it not to be set on
rendering. Subsequent edits, then, would likely cause the term to be
null and cause all kinds of fun errors. This commit fixes that issue.