Give returning instructors the option to start from an empty timeline.
Add an additional wizard option for instructors who already have a course in the system, when they are creating another, so that they can start from an empty timeline.

I accomplish this by adding a new yaml file that can get combined with the wizard_index, adding the additional wizard option only if the instructor is returning. If we need to add more conditionals to the available assignment options later on, it will be better to refactor TimelineController#wizard_index into a separate manager that checks conditions specified in the main index, similarly to the dependency handling in the wizard content files.