Improve WikiEdits behavior when removing user from course
When a user gets removed from a course, we don't necessarily want to remove the assignment templates for that user's assignments, because classmates may still be assigned to that same article. It's safer to just run WikiEdits.update_assignments after that user's assignments have been destroyed.

More broadly, the behavior of WikiEdits.update_assignments is too complex, because it can either update all assignments for a course or it can update just some based on assignment data passed to it, and it can be in 'delete' mode or normal mode. It will be better to just always do updates based on the course and the current_user, meaning that any user updating the assignments for a course will post updates to all relevant pages every time. The 'delete' part gets tricky, though.

Resolves #431