[multiline] Disable course save if necessary params are absent
Courses were somehow getting saved with a null slug. I still don't know
why that was, or how that happened. But when that occurred, the check
for the course on course creation looks up params[:id], which is nil, so
it looked for a course with a null slug. Because there was a course with
a null slug, it found one, thus circumventing the check that is supposed
to give a user permission to create a course if
`Course.find_by_slug(params[:id])` returns nil. This commit adds a
`before_save` validation that will fail if any of the necessary params
are absent, which should prevent courses getting saved with a null slug
in the future.