Add interface for new faster pageview API
This new API directly from Wikimedia dramatically cuts down the time it takes to get pageview data. It's still experimental as of now, and only has data going back to October 1 2015. Eventually it will be backfilled to May 2015, but no further. Once it is stable and out of experimental status, we can build out this class and replace Grok entirely.

In the short term, it dramatically increases the utility of article_finder. Now the bottleneck is ORES... which we can solve by threading our requests.