Parallel builds on travis: features / everything else
Set up a build matrix to run all the feature specs separately from the rest of the specs. Set the deployment script to only run for one build of the matrix.

This means that deployment will still happen if feature specs fail, since they are in a different build. But we should only be pushing to staging or production after the build has already passed on master or another branch, so the upshot will be that the build environment for deployment is consistent, and deployment happens faster.

This should hopefully cut down on the time it takes to run the specs on travis.