Refactor CourseMeetingsManager to reduce duplicate calculations
All the uses of CourseMeetingsManager require the week_meetings, so we should just calculate that once each time we need a new CourseMeetingsManager.

We could improve things further by refactoring the way the TrainingModuleDueDateManager interacts with CourseMeetingsManager, so that it doesn't create a new one for every training module assigned for a course.