Refetch the plagiarism report urls on each updates.
These urls expire, so we can work around that by getting it anew from the plagiabot API each time the update process runs.

Resolves #493

It will be better in the long run to fetch these urls as-needed, but as long as we don't have a huge number of flagged plagiarism revisions to deal with, it's not a big deal.