Add link to refresh plagiarism reports on demand.
The report urls for iThenticate, provided via the plagiabot API, expire after a short time — possibly every ~20 minutes. This is a somewhat clumsy solution, but it should make the reports accessible. It simply provides a link that triggers a refresh of the data, and then redirects back to the plagiarism feed.

Resolves #493