[multiline] Stub ServerActions method to avoid AJAX request
Prior to this commit, all Mocha tests were passing individually, but
failing as a suite. This was because `ServerActions` was calling methods
that require an AJAX request, which failed, adding a notification to the
DOM (thus making the assertions on the number of notifications
incorrect). This commit adds Sinon.js and stubs the call from
ServerActions that fires a request, making sure the original is not
called, and preserving the intended functionality of the Notifications
spec.