[multiline] Fix issue with timeline spanning across years
`moment(date).weeks()` returns the week of the year, so if the course
timeline spans across a year, the subtraction of the end (e.g., 45) minus
the start (e.g., 1) gets a large negative number, and thus when compared
with the number of weeks in the course, is smaller, so `timeline_full`
returns false, when in fact it may be. This commit looks at the week
difference between the end and start dates (accounting for them spanning
across a year) and compares that not with the number of `week` objects,
but with the week_components to check if there is room in the timeline.