[multiline] When [] is passed for training_module_ids, set the attr accordingly
When the API sends `[]` for training_module_ids for a block, ActiveRecord
sets it to `nil` because of the possiblity for unsafe query generation
(see
http://guides.rubyonrails.org/security.html#unsafe-query-generation).
Unfortunately, then the controller doesn't accept that as a valid param
in the whitelist (because it's expecting an array), so it ignores the
param. (See [this github
issue](https://github.com/rails/rails/issues/13766#issuecomment-32730118)
for more details.)

This PR conditionally adds whitelisting for `[]` for a block's
training_module_ids, and thus allows clearing them out.

Closes #525.