Switch to WikiPageviews API to collect daily pageviews
This should make the pageview data more accurate, and the pageview update process more reliable.

Also, it tweaks the ArticlesCourses.update_from_revisions to run against all revisions even if some revisions don't have Users in the system. This is relevant for article_finder, which pulls in Articles and Revisions but not the Users who made those revisions.