Don't mark articles as deleted if Replica is down.
The ArticleImporter.update_article_status routine assumed that any article that didn't get returned from a request to Replica must be deleted. This meant that if Replica's webservice was not available and every request failed, the system would update every single article as deleted. There's still room for improvement; this is just a quick patch to work around the problem, but it would be better if relied on a positive confirmation that an article was deleted, for example by querying the deletion logs.

Resolves #583