Delete pid files if the processes aren't running
When the system crashes hard, the update pid files may be left behind even though the processes died. This will delete them and allow updates to continue, if the pid files do not have corresponding processes.

Also, use a separate file for the 10 minute pauses instead of combining those with the manually-triggered batch_pause file.

Resolves #288 (probably)