On `capistrano deploy:rollback`, do not overwrite assets
The build and upload asset steps of capistrano's deployment script have also been applied to rollback, meaning that the compiled assets were not being rolled back but instead compiled and uploaded anew... and deployed against the older version of the ruby code. I couldn't find a better way to switch off asset steps for rollback, and this method seems to work.

Resolves #599