Fix course overview spec.
Now that the dates and meeting days are being calculated more accurately in ThisWeek, the meeting days would vary based on when the test was run. Setting dates to a fixed ones far in the future solves this and accurately tests more of the intended behavior.

(Also, remove an unused method from ThisWeek compontent.)