Improve error handling for MediaWiki queries
An update rake task failed because of a commons query that returned `MediawikiApi::ApiError: Invalid value '' for user parameter ucuser (ucbaduser_ucuser)`. This resulted in `nil` propogating until it caused an error in Commons.get_uploads. This doesn't directly fix that problem, but it will help gather enough info to make it possible to piece it together in the future.