Print Greeked passcode
Differentiate between a course with no passcode, and one in which the passcode
is hidden from your role, by displaying a bunch of asterisks if a course is
passcode-protected but you do not have permissions to see the code.

This might help to set expectations for people hoping to join the course.

Also provides i18n for the field.