Rename User.wiki_id to User.username
The loose ends of this are that the Replica endpoint still had 'wiki_id' in its output, and a few update methods rely on that matching up with an attribute on User. We can first push this update to all dashboard instances that rely on the replica endpoint, then we can change the replica endpoint to return 'username', and then we can remove the alias_attribute on User.