Make AssignedArticleImporter wiki-specific
Also, fix some wiki attribute stuff on Assignment and Article

For some reason, the before_validation does not run when an Article gets created in a spec. The only way I could ensure that the defaults get set is to add an after_initialize callback. Maybe that's just an issue with factory_girl. Investigation on production is needed to see where wiki-related attributes are missing.