Make Replica return `nil` more reliably if query fails
Likely resolves #734

We can count on 'success': true in the response from the replica endpoint, if the query is successful.

One loose end with the Replica interface is that .get_revisions behaves differently; it returns an empty hash instead of nil if the query fails. For consistency, we might want to change it to return nil and handle that eventuality where get_revisions is called.