Possible fix for intermittent VCR no-cassette error on travis
On vagrant with xvfb for the feature specs, I was able to replicate the intermittent VCR error that we see frequently on travis.

Also, ensure hot_loading doesn't interfere with feature specs. This means `gulp build` or `gulp` must be run before feature tests are started, if `gulp hot-dev` was being used.