Fix assignment store setup
An 'id' attribute is necessary for populating the store will all the records. However, new items in the store get a new id, and we don't want to pass the new ids back to the server for deletion requests. So we duplicate the id as 'assignment_id' as well, and then everything plays nicely.