Use markdown editor for all survey question types
Trix is much less flexible than markdown, and we want to be able to embed a video in a Text pseudo-question. This was possible with all other question types, but not with Text unless you created it first as a different type and then switched to Text type.

Fixing bugs solely by removing code for the win!