Clean up date props
The timeline date limits were broken anyway, since they were using @props.weeks instead of @props.weeks.length. But I'm not sure what the optimal behavior is.

In FormPanel, there should be no weeks yet anyway, so the user should probably be free to use any timeline range that is consistent with the course start/end dates.

In Meetings, one option would be to keep the user from shrinking the course below the number of weeks, but that has the added complication of week boundaries: a 7 day course from Wednesday to Tuesday spans two calendar weeks and would have two weeks in @props.weeks.

So, for now, we'll just remove the limitations altogether, except to ensure that the course and timeline dates constrain each other.