Improve handling of timeline save failure
If a timeline save fails because timeline content changed on the server in the meantime — typically because of another edit that was made in another browser — it would keep failing on subsequent saves as well, without refetching the current timeline data from the server.

This adds a detailed error message and also refreshes the course and timeline data, so that users can more easily recover from timeline save failures.

Partially addresses #950