Tweak spec so it passes in all time zones
This started failing in negative UTC offset timezones, but not travis, after we implementing start/end times and timezones.

This just works around it, rather that testing the precise behavior of timezones and meeting days. The inconsistencies should go away if/when we switch to using local start/end time for all course types.