Rework gulp and asset tagging to fix RTL issues
In order to make relative urls in RTL stylesheets work, we need to serve them from the same directory as the default stylesheets.

This saves the stylesheets in the same directory, with the same revision hash, but with an "rtl-" prefix. Corrsponding tweaks to ApplicationHelper make sure the correct files are served for both hot_loading and default environments.