Prevent mistaken marking of articles as deleted
Some articles have recently been marked as deleted by mistake recently. I think this was happening because Replica went down in the middle of processing the articles that didn't turn up in a search by id, but may still exist with a different id. If so, this should prevent it.