Convert CourseClonedModal to JSX
In previous attempts, this was not working after conversion in Poltergeist (but was fine in Selenium). It turns out that this is because of the use of [a,b].includes(x), which doesn't work as expected in Poltergeist. Elsewhere in the codebase, we use lodash for that — _.includes([a,b], x) — which works fine.