Switch to current travis image with updated MySQL
With MySQL 5.6, the TrainingModuleDueDateManager spec behaves differently from in 5.5, presumably because of the order in which records get created and numbered in combination with the way weeks of a course get reorded. The changes to the spec sort that issue out by avoiding use of some of the complex 'let' setup.