Fix relative link issues with ArticleViewer
ArticleViewer inserts html from the mediawiki RESTbase API. This includes a 'base' element that lets the article styling work like on Wikipedia, but it also sets the mediawiki server as the base url for the rest of the dashboard page.

'ClickOnOutside' intercepts the clicks so that they end up going to the right place, but using absolute urls in the nav means they will also be correct on mouseover. The feedback button link is fixed directly in the ArticleViewer component.

Styling is still disrupted, but this should make everything basically functional.