Make sure chat iframe shows after switching tabs
Currently, we log in to Rocket.Chat every time we load this tab. It would be nice to maintain the iframe's state even when it isn't being shown. Perhaps this would involve putting it at the top level of course.jsx and just hide it whenever it's not in use.