Add more threads for main update bottlenecks
Update times are getting very long, especially on P&E Dashboard. The biggest contributors to the bottleneck are the import of new revisions through Replica, and the database-intensive calculations in the CoursesUsers cache update process.

This bumps up the threads for CoursesUsers cache updates to 5. These updates are limited by the number of available cores, but both P&E and Wiki Ed production currently have enough cores for this to make a difference (8 and 6 respectively).

It bumps the concurrency limit for Replica queries to 10. The revision importation is mainly limited by the slowness of queries on Replica, but there are no known problems related to hitting replica concurrently with many queries at once.