Extend existing validations to all wikisource
These changes are 2 fold, adjustments to the validation & creating
methods for wikisource and some refactoring of the specs for speed and
readability.

Wikisource changes

Typically a wiki page has a language support, but in some edge cases a
given page will support many languages. This has led to specific logic
around wikidata and wikisource. They are both multilingual in nature but
wikisource can SOMETIMES have a language assigned depending on what
resource within wikisource is being worked on.

I consider these changes an intermittent step. Long term we want a way
to ask the services what languages they support for a given project and
do validation that way. Until that work is completed, this code will
  solve our current use case.

Refactoring of the spec

In some case, a create statement was not needed, so we can save time by
using FactoryGirl#build. Other places I was able to reduce the lines of
code by using object equality checks instead of going through property
comparisons. I also did my best to add some context descriptions to try
and help the documentation quality a little.