Extend API to handle wikisource projects
Includes a small change to the 'assignments#create' workflow, when
checking for missing pages in the API it now ensures that the 'missing'
key has a value.

'www' is now a standard language symbol that is intended to mean a
multilingual page.

Model checks were removed from the feature spec, they were causing
instability in the test suite because feature specs are run in a
separate thread than the main rails execution which creates race
conditions. Confidence is still high here because of the controller
specs and the process involved in properly rendering assignments after a
page refresh indicates a correct back-end response.