Refactor block date calculations
This takes the date calculation for a single block out of TrainingModuleDueDateManager and makes it accessible directly from the block, and also adds a method to get the start date for the block, representing when the activity in the block starts rather than ends.

This will let us more easily calculate the dates of important blocks.