Add stubbing of wiki validation in specs
Now that we'll be making a web request before creating a new wiki, this requires handling the requests throughout our test suite whenever new wikis are created.

The strategy here is to stub all the requests that happen throughout most of the tests, and use VCR for the ones testing the Wiki model itself. This is not entirely satisfying, especially because it will require adding that stub method in any new specs that involve creating a wiki and we'll need to maintain the list of validated wikis. I can't think of anything better at the moment, though.