Upgrade to react-trix 0.2.1 and trix 0.10.1
react-trix has a new interface: instead of sending events to its onChange handler, it sends the html and text representation of the content.

We must package this into an event with a target and value to match the expected input for the InputMixin's change handler.

Hat tip to dstpierre for helping me figure this out at https://github.com/dstpierre/react-trix/issues/7