Test for regression of broken CourseNavbar links
This is one way to access and test the props being passed to child components. Because of quirks with how react-router's Link element is handled, the href's aren't available in the test environment unless we add some scaffolding for a Router to wrap it, but with shallow rendering and accessing the child nodes, we can verify that the right prop is passed to Link (without relying the behavior of the Link component itself).

This covers the bug fixed in 6d2c7687b98cf306e95ded6316694842e7c57cf7