Convert GetHelpButton/AlertsStore to use Redux
This involves the first use of API actions in connection with Redux. Thunk middleware appears to be the standard way to doing this, so that multistep actions with side-effects can be used in combination with the Redux store.

Here we replace the AlertsStore with a reducer, rewrite AlertActions into redux action creators, and introduce a shim for the API_FAIL action to use Flux still, since that isn't yet handled by the Redux store.