Use tools.wmflabs.org instead of whocolor.net for ArticleViewer
The whocolor.net security certs have expired, but it's just the PHP scripts from https://github.com/jurkan/whoCOLOR

The source wikiwho data comes from another server (with a hardcoded IP), so I've copied the scripts to Tool Labs and they still work (with a tweak to use secure Wikipedia endpoints). This makes for a simple workaround for the certs problems.