Avoid unecessary work in calculating trained_count for course
This should eliminate the bottleneck for updating Course caches, which was mainly due to the unnecessary creation of many extra TrainingProgressManager objects and their slow due date calculations. The TrainingProgressManager was only used to check for completed status, but was doing a lot of extra work and extra queries during initialization that aren't necessary.