Fix Javascript to work with new Haml 5.0 escape behavior
The interpolated Ruby in a Haml :javascript filter now gets HTML escaped, breaking the way it was being used in the shared _head template to pass the set of valid languages and projects.

The fix is to pass it as an HTML dataset instead, and pull that into the Javascript afterwards. I followed the example of https://github.com/sferik/rails_admin/pull/2870