Use MetaWiki instead of Replica to update users
Since we no longer need to check 'trained' status of users, we can easily use Meta directly to fetch info about users, in particular global_id.

This will also allows us to fetch users' Meta registration date, which we will be able to use to determine which users are newbies.