Treat 404 from WikiPageviews as 0 views
The WikiPageviews API has the quirk that it returns a 404 when it cannot find any views for an article in the requested date range. In order to set the correct view number of 0 (instead of nil) and avoid repeatedly trying to fetch the same (lack of) data, WikiPageviews needs to interpret the 404 response as 0 views.