Adjust the instructions in CourseCreator depending on the step
Instead of just a single message, show one relevant to the current state of the CourseCreator, whether choosing a clone or filling out the course form.

Also, refactor and rename things to make the statefullness of the component more clear: it basically takes one of three forms, but this wasn't very clear from the code.