Adjust 'Continue improving' block, add script to update old blocks
We want to include the images training and an images-related note in the 'Continue improving your article' timeline block. The timeline change will put that in place for future courses, and the script can be run to update old courses so that when they are cloned they will have the updated content and training.