Create a real CSV builder for before/after ORES scores
This adds a (hidden) route to generate a CSV of all the edited articles for a campaign, including the characters added and the ORES wp10 scores before and after.

It assumes that all the necessary wp10_previous scores have already been imported; only the manual ores_changes script does that systematically for all the articles edited in a campaign, so this would need to be done before pulling any data.