Add importing of previous wp10 scores to the daily update
With concurrency, this is fast enough that we should be able to run it as part of the daily update.

Eventually, we may want to improve how non-existint previous revisions are handled so that we do not re-query for their data every day. But it should be okay for now.