Switch to new student user and talk templates
These new templates include the 'slug' param, which allows for working around the limitation on new accounts posting external links, by using an internal link when the page is first saved and only switching to external after a subsequent edit to create the user's sandbox has been made.

This should be backwards compatible with any of the templates currently in use, as it just adds a new parameter that won't be used by templates that haven't been updated.