Integrate course level with Salesforce
This will pass the course level only if its set, so that we don't overwrite courses where we already have a value in Salesforce but not in the dashboard.

The options in CourseLevelSelector are adjusted to match the ones we use in Salesforce (although it's currently a text field that will accept standard values as well, on the Salesforce end).