Deleted revisions should not contribute to stats
For cases where an article is deleted and immediately redirected, the dashboard knows that all the revisions are deleted but since the article itself still exists, it doesn't treat them any differently. This fixes the behavior of ArticlesCourses and CoursesUsers data caching so that the stats for courses and campaigns don't include deleted edits.