Rework cloned course saving flow to avoid uncaught error
For some reason, there was an uncaught error breaking the spec, which happened upon redirect, after I added the babel-polyfill to main.js.

This refactors the submission flow to use the same redirection pattern that we use for retitling existing courses, and also extends the cloning spec to first submit an already existing course slug, then fixing it and saving again.