Avoid database query in pageview import
There was a database pool timeout error in P&E Dashboard production during the view import cycle, which timed out trying to query the wiki from the article in WikiPageviews.

By including the wiki in the original query for all articles that need view updates, we should be able to avoid that problem.