Refactor Confirm to use redux
This completely rewrites the way the react 'Confirm' component works: intead of inserting the component where a confirmation dialog is needed, it puts a single instance in the Course component, and only takes an redux action to activate. The onConfirm callback is passed to the action to become part of the redux state, along with the other props for the dialog: the message, additional explantory text, and whether to show an input box.

This should make it much less cumbersome to use the react component for confirmation dialogs, so we can replace the native javascript confirms.