Increase Dalli cache max file size
Recently, training modules stopped working on outreachdashboard.wmflabs.org. This appears to have been caused by the training content and translations having grown beyond the 1M default file limit of memcached / Dalli.

I've now changed the memcached config on the production server to allow larger files. In combination with this change, it should restore functionality to the training system.

In the longer term, we need a better storage strategy for training content. Moving the trainings into the database is probably the right solution.