Rename vendor.js to jquery-uls.js
This is mainly being done to prevent errors with users who have a cached copy of vendor.js. The older version of that file includes trix-editor, which is now part of the main.js file. This causes an error and breaks the dashboard for users with the old (non-fingerprinted) vendor.js.

Since the vendor javascript file now only has the scripts needed for the universal language selector, this new name makes it clearer what it is used for.