Refetch course from server after changing campaigns, fix spec
With the new data flow, the 'submitted' and 'published' state of courses would not change after adding or removing campaigns. To work around that, we can fetch the course again from the server after updating the campaigns.

The feature specs for this behavior also go muddled a little bit; this sorts them out.