Add 'withdrawn' attribute to course, use it for 'approved' status
This will let us mark a course as 'withdrawn', and thereby deactivate it without removing it from a campaign, since we use campaigns in sync with Salesforce to track and project metrics for all courses including the withdrawn ones.

This should also prevent the system from sending emails to instructors for withdrawn courses.