Fix training-related specs
I couldn't figure out how to just load the training slides into the database once for the entire test suite, so it happens via `TrainingSlide.load` wherever the specs interact with training slide content. It would be nice to find a better way that would treat TrainingSlide.load as part of the pre-test database setup alongside the fixture, so that it would just roll back to a state that included the slide records after each test.