Convert TrainingSlide into an ActiveRecord model (#1732)
Instead of storing training slides in memory, we're moving them into the database. This should fix the reliability issues with the wiki training slides, and make it easy to update individual modules or slides with new wiki content and translations, without the potential for breaking other trainings or getting into a broken state due to rate-limiting on the Wikipedia end of things.

Eventually, we'll probably want to convert TrainingModule and TrainingLibrary to ActiveRecord as well, but for now, this does the minimum necessary to support database-backed TrainingSlide objects.

We'll probably want to institute caching for the training views soon, which should avoid most of the performance hit of moving to ActiveRecord.