Convert AssignmentStore to redux
This gets rid of another StockStore and moves the assignments data flow to redux.

Next step will hook in the server-side data flow more fully, doing 'addAssignment' based on data that gets returned from the server so that we always have accurate IDs on the client. Then, we can implement a mass-new-assignment input for AvailableArticles that can loop through input and call addAssignment on each one.