Remove Article Finder feature flag and results route
With the client-side prototype, there's no longer any reason to keep this disabled in production. Just to be safe, this removes the #results route and action.

CategoryImporter still has a lot of now-dead code that was only used for ArticleFinder, but it's also the only code that uses AverageViewsImporter, which we will probably want to save for #1809