Add a library for Wikipedia Fellows trainings, option to hide
This adds an index page for the Fellows training modules, and adds a new option used by the Fellows library to exclude it from the top-level training index. The Fellows program is pretty niche, so we don't want to clutter the training index with it, but having a library means then breadcrumbs won't point back to a nonexistent page.