Add fallback for four-byte unicode characters in article titles
This is a temporary fix for MySQL errors caused by four-byte unicode characters. The database can currently only handle three-byte unicode characters, so some emoji and other characters cause errors when articles are imported.

This is a workaround that saves the titles in URL encoding when that error occurs, so that they will still basically work except that the titles won't be correct.

Addresses #1744