Move handling of duplicate assignments to the server
The duplicate assignment check on the client did not account for language and project, because for the home wiki these are not set. This prevented assigning the same title for multiple languages.

Moving the error duplicate handling completely to the server, and including the relevant error message, solves this bug.

Resolves https://phabricator.wikimedia.org/T197764