Remove no-longer-needed code for replacing assignment templates
Since the switch to using the templates config, the 'old' template being checked for was actually the same as the 'new' one. It originally replaced an older template with a different name, to avoid double-tagging courses after the name change. That's no longer necessary because any courses that used the old name are long-since ended, and it wouldn't have worked since the switch to the templates config anyway.