Set dashboard_url ENV VAR for test environment
Added the dashboard_url ENV VAR to the test environment config file.

Some tests were failing as the test environment is expecting a different
endpoint than what is set in application.yml. All environment variables
that tests rely on are consolidated to the config file as the
application.yml and the test config file may have different settings.