Add 'text-align: left' to remove the blue color to the left of the search input.
Add 'display: block' because the class '.top-nav__faq-search', which is previously defined on line 105, is applying 'display: none' due to line 138.

Use class '.icon-search' to add space to the left of the search icon, which only effects this instance of the '.icon-search' since it's in the media query.

All JavaScript tests passing
Not able to run Ruby tests